requirements.txt
Flask
pillow
numpy
requests
watson-machine-learning-client